git-svn-id: http://einstein/svn/firmware/FIR-00251-Proc/trunk@27403 3e98269d-ff51-e449-abca-491a6567d830

+++ src/sw/Calibration.h
@@ -223,7 +223,7 @@ struct calibCollectionInfoStruct {
    uint32_t POSIXTime;                 /**< Collection file generation date and time. */
    CalibrationCollectionType_t CollectionType; /**< Collection type */
    calibrationType_t CalibrationType;  /**< Calibration type */
-   uint8_t SensorID;                   /**< Sensor ID */
+   uint16_t SensorID;                   /**< Sensor ID */
    uint8_t IntegrationMode;            /**< Integration mode */
    uint8_t SensorWellDepth;            /**< Sensor well depth */
    uint8_t PixelDataResolution;        /**< Effective size in bits of a pixel of the image */
